Modify map methods
